Allow to read env from dotenv file
