+++ bake/compose.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"os"
 	"reflect"
-	"strings"
 
 	"github.com/docker/cli/cli/compose/loader"
 	composetypes "github.com/docker/cli/cli/compose/types"
@@ -15,26 +14,20 @@ func parseCompose(dt []byte) (*composetypes.Config, error) {
 	if err != nil {
 		return nil, err
 	}
+
+	envs, err := readEnv()
+	if err != nil {
+		return nil, err
+	}
+
 	return loader.Load(composetypes.ConfigDetails{
 		ConfigFiles: []composetypes.ConfigFile{
 			{
 				Config: parsed,
 			},
 		},
-		Environment: envMap(os.Environ()),
+		Environment: envs,
 	})
-}
-
-func envMap(env []string) map[string]string {
-	result := make(map[string]string, len(env))
-	for _, s := range env {
-		kv := strings.SplitN(s, "=", 2)
-		if len(kv) != 2 {
-			continue
-		}
-		result[kv[0]] = kv[1]
-	}
-	return result
 }
 
 func ParseCompose(dt []byte) (*Config, error) {
